developed: user, card

+++ src/utils/Api.js
@@ -96,25 +96,27 @@
                  }
              })
      }
-     likeCard(id) {
-         const cardLikeURL = (this._cardsURL + `/likes/${id}`);
-         return fetch(cardLikeURL, {
-                 method: 'PUT',
-                 headers: {
-                     authorization: this._token
-                 }
-             })
-             .then(res => {
-                 if (res.ok) {
-                     return res.json();
-                 } else {
-                     return Promise.reject('Ошибка связи с сервером');
-                 }
-             })
+    //  likeCard(id) {
+    //      const cardLikeURL = (this._cardsURL + `/likes/${id}`);
+    //      return fetch(cardLikeURL, {
+    //              method: 'PUT',
+    //              headers: {
+    //                  authorization: this._token
+    //              }
+    //          })
+    //          .then(res => {
+    //              if (res.ok) {
+    //                  return res.json();
+    //              } else {
+    //                  return Promise.reject('Ошибка связи с сервером');
+    //              }
+    //          })
 
-     }
-     dislikeCard(id) {
-         const cardLikeURL = (this._cardsURL + `/likes/${id}`);
+    //  }
+     changeLikeCardStatus(id, isLiked){
+       const cardLikeURL = (this._cardsURL + `/likes/${id}`);
+       console.log(cardLikeURL);
+      if(isLiked){
          return fetch(cardLikeURL, {
                  method: 'DELETE',
                  headers: {
@@ -128,7 +130,38 @@
                      return Promise.reject('Ошибка связи с сервером');
                  }
              })
+     }else{
+      return fetch(cardLikeURL, {
+              method: 'PUT',
+              headers: {
+                  authorization: this._token
+              }
+          })
+          .then(res => {
+              if (res.ok) {
+                  return res.json();
+              } else {
+                  return Promise.reject('Ошибка связи с сервером');
+              }
+          })
+      }
      }
+    //  dislikeCard(id) {
+    //      const cardLikeURL = (this._cardsURL + `/likes/${id}`);
+    //      return fetch(cardLikeURL, {
+    //              method: 'DELETE',
+    //              headers: {
+    //                  authorization: this._token
+    //              }
+    //          })
+    //          .then(res => {
+    //              if (res.ok) {
+    //                  return res.json();
+    //              } else {
+    //                  return Promise.reject('Ошибка связи с сервером');
+    //              }
+    //          })
+    //  }
      avatarUpload(url) {
          return fetch((this._userURL + `/avatar`), {
                  method: 'PATCH',
